feat: add allNetworkByloop
